# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Legacy Link Backend API",
    "version": "0.6.0",
    "description": "API Legacy Link: token đọc cho GET, token ghi cho POST; timestamp đo là epoch milliseconds, ID DB là chuỗi. C16 bổ sung profile version, preview, lịch sử thao tác PostgreSQL, metrics, freshness. Xem deploy/README.md và docs/SO-TAY-HIEU-DU-AN-LEGACY-LINK.md. Rate limit mỗi socket IP/process: 300 GET và 30 thao tác/phút; 429 kèm Retry-After. Thêm deliveryHealth và diagnostics.delivery để quan sát queue ESP32; không cần endpoint riêng. AI: POST /ai/chat và /ai/query chỉ đọc, chấp nhận token đọc; Gemini chỉ chọn tool. UNDERHEAT dùng lowAlarm riêng, yêu cầu đơn vị nhiệt độ."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/machines": {
      "get": {
        "summary": "Danh sách máy đã đăng ký",
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Machine"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu token hoặc token không hợp lệ"
          },
          "403": {
            "description": "Không đủ quyền thao tác hoặc Origin không được phép"
          },
          "429": {
            "description": "Quá số request; chờ Retry-After (giây)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/alarms": {
      "get": {
        "summary": "Lọc cảnh báo trên toàn hệ thống",
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmPage"
                }
              }
            }
          },
          "400": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu token hoặc token không hợp lệ"
          },
          "403": {
            "description": "Không đủ quyền thao tác hoặc Origin không được phép"
          },
          "429": {
            "description": "Quá số request; chờ Retry-After (giây)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "low",
                "medium",
                "high",
                "critical"
              ]
            }
          },
          {
            "name": "acknowledged",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "deviceId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/gateways": {
      "get": {
        "summary": "Danh sách ESP32 đang báo trạng thái",
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Gateway"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu token hoặc token không hợp lệ"
          },
          "403": {
            "description": "Không đủ quyền thao tác hoặc Origin không được phép"
          },
          "429": {
            "description": "Quá số request; chờ Retry-After (giây)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/catalog": {
      "get": {
        "summary": "Đọc cấu hình máy; catalog cũ có thể dùng các trường alarm dạng phẳng",
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu token hoặc token không hợp lệ"
          },
          "403": {
            "description": "Không đủ quyền thao tác hoặc Origin không được phép"
          },
          "429": {
            "description": "Quá số request; chờ Retry-After (giây)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "deviceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/uptime": {
      "get": {
        "summary": "Lịch sử chạy consumer; không phải bằng chứng chắc chắn mất mẫu",
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Run"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Lỗi hoặc tạm thời chưa phục vụ được",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu token hoặc token không hợp lệ"
          },
          "403": {
            "description": "Không đủ quyền thao tác hoặc Origin không được phép"
          },
          "429": {
            "description": "Quá số request; chờ Retry-After (giây)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
    "/profiles": {
      "get": {
        "summary": "Danh sách bản mới nhất của từng profile",
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfileSummary"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi yêu cầu hoặc dịch vụ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Lỗi yêu cầu hoặc dịch vụ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Lỗi yêu cầu hoặc dịch vụ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Lỗi yêu cầu hoặc dịch vụ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Lỗi yêu cầu hoặc dịch vụ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Lỗi yêu cầu hoặc dịch vụ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Lỗi yêu cầu hoặc dịch vụ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Lỗi yêu cầu hoặc dịch vụ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 100
            }
          }
        ]
      }
    },
...
    "/ai/chat": {
      "post": {
        "summary": "Câu hỏi tiếng Việt: Gemini chọn công cụ; backend xác minh dữ liệu",
        "description": "Chấp nhận read hoặc write bearer token. Chỉ gửi question hoặc action. Context giữ ID kết quả 10 phút, tối đa 200 phiên/process. Model tối đa 5 câu/phút/process, 1 đồng thời; 1 request/câu, không retry. Timeout provider 12s, tổng tools 8s. Không có key/hết quota thì fallback minh bạch. Không thực thi SQL, URL hoặc điều khiển máy.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "conversationId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "question": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1000
                  },
                  "action": {
                    "type": "string",
                    "enum": [
                      "overheat",
                      "underheat",
                      "offline",
                      "summary",
                      "alerts"
                    ]
                  }
                },
                "oneOf": [
                  {
                    "required": [
                      "question"
                    ]
                  },
                  {
                    "required": [
                      "action"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dữ liệu từ database; mode=rules hoặc gemini; không bịa số đo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CopilotResponse"
                }
              }
            }
          },
          "400": {
            "description": "Sai tham số"
          },
          "401": {
            "description": "Thiếu hoặc sai token"
          },
          "403": {
            "description": "Origin/phiên không được phép"
          },
          "404": {
            "description": "Máy không tồn tại"
          },
          "429": {
            "description": "Rate limit HTTP"
          },
          "503": {
            "description": "Nguồn dữ liệu không khả dụng"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AlarmPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Alarm"
            }
          },
          "from": {
            "type": "integer",
            "format": "int64"
          },
          "to": {
            "type": "integer",
            "format": "int64"
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "from",
          "to",
          "nextCursor"
        ]
      },
...
      "ProfileSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "revision": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token đọc hoặc ghi do người vận hành cấp; POST yêu cầu token ghi."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}